Fix: file이름 수정, 테이블 config파일 생성
